+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - New #211: Add `receiverUserId` and `callbackQueryId` parameters to `SendMessage`, `SendAnimation`, `SendAudio`,
   `SendDocument`, `SendPhoto`, `SendSticker`, `SendVideo`, `SendVideoNote`, `SendVoice`, `SendContact`,
   `SendLocation` and `SendVenue` methods.
+- New #211: Add `ephemeralMessageId` field to `ReplyParameters` type.
+- Chg #211: Make `messageId` field of `ReplyParameters` type optional.
 - Enh #211: `RichText` types, `RichBlockCaption`, `RichBlockTableCell` and `Location` now support serialization
   back into a request array via `toRequestArray()`.